+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 ## vNEXT
-- Remove KYC tests stage from Jenkinsfile. (#51)
+- Change MNEMONIC var name for production & clean Hardhat file. (#53)
 - Update headers & reformat Poco1 contracts. (#52)
+- Remove KYC tests stage from Jenkinsfile. (#51)
 
 ## v5.5.0 - PoCo Boost
 - Migrate to Hardhat: